Make a navigation utils file
