+++ web/src/js/navigation/__tests__/navigation.test.js
@@ -22,7 +22,7 @@ afterEach(() => {
 
 describe('goTo', () => {
   it('calls history.push as expected', () => {
-    const { goTo } = require('../navigation')
+    const { goTo } = require('js/navigation/navigation')
     goTo('/some/path/')
     expect(mockBrowserHistory.push).toHaveBeenCalledWith({
       pathname: '/some/path/',
@@ -31,7 +31,7 @@ describe('goTo', () => {
   })
 
   it('calls history.push with query parameters', () => {
-    const { goTo } = require('../navigation')
+    const { goTo } = require('js/navigation/navigation')
     goTo('/some/path/', { someParam: 'abc', foo: 'blah' })
     expect(mockBrowserHistory.push).toHaveBeenCalledWith({
       pathname: '/some/path/',
@@ -42,7 +42,7 @@ describe('goTo', () => {
 
 describe('replaceUrl', () => {
   it('calls history.replace as expected', () => {
-    const { replaceUrl } = require('../navigation')
+    const { replaceUrl } = require('js/navigation/navigation')
     replaceUrl('/some/path/')
     expect(mockBrowserHistory.replace).toHaveBeenCalledWith({
       pathname: '/some/path/',
@@ -51,7 +51,7 @@ describe('replaceUrl', () => {
   })
 
   it('calls history.replace with query parameters', () => {
-    const { replaceUrl } = require('../navigation')
+    const { replaceUrl } = require('js/navigation/navigation')
     replaceUrl('/some/path/', { someParam: 'abc', foo: 'blah' })
     expect(mockBrowserHistory.replace).toHaveBeenCalledWith({
       pathname: '/some/path/',
@@ -65,7 +65,7 @@ describe('absoluteUrl', () => {
     // Set the domain env var
     process.env.REACT_APP_WEBSITE_DOMAIN = 'some.example.com'
 
-    const absoluteUrl = require('../navigation').absoluteUrl
+    const absoluteUrl = require('js/navigation/navigation').absoluteUrl
     expect(absoluteUrl('/')).toBe('https://some.example.com/')
     expect(absoluteUrl('/blah/')).toBe('https://some.example.com/blah/')
   })
@@ -74,272 +74,7 @@ describe('absoluteUrl', () => {
     // Set the domain env var
     process.env.REACT_APP_WEBSITE_DOMAIN = 'some.example.com'
 
-    const absoluteUrl = require('../navigation').absoluteUrl
+    const absoluteUrl = require('js/navigation/navigation').absoluteUrl
     expect(absoluteUrl('https://foo.com/blah/')).toBe('https://foo.com/blah/')
-  })
-})
-
-describe('isURLForDifferentApp', () => {
-  it('[absolute URL] considers the apps the same if the URL is exactly the same', () => {
-    const { isURLForDifferentApp } = require('../navigation')
-    setWindowLocation({
-      host: 'example.com',
-      hostname: 'example.com',
-      href: 'https://example.com/blah/path/',
-      origin: 'https://example.com',
-      pathname: '/blah/path/',
-      port: '',
-      protocol: 'https:',
-      search: '',
-    })
-    expect(isURLForDifferentApp('https://example.com/blah/path/')).toBe(false)
-  })
-
-  it('[absolute URL] considers the apps the same if the URLs are pages on the same "homepage" app (not part of the "newtab" or "search" subpaths)', () => {
-    const { isURLForDifferentApp } = require('../navigation')
-    setWindowLocation({
-      host: 'example.com',
-      hostname: 'example.com',
-      href: 'https://example.com/blah/path/',
-      origin: 'https://example.com',
-      pathname: '/blah/path/',
-      port: '',
-      protocol: 'https:',
-      search: '',
-    })
-    expect(isURLForDifferentApp('https://example.com/some/thing/')).toBe(false)
-  })
-
-  it('[absolute URL] considers the apps different if the hostname is different', () => {
-    const { isURLForDifferentApp } = require('../navigation')
-    setWindowLocation({
-      host: 'example.com',
-      hostname: 'example.com',
-      href: 'https://example.com/blah/path/',
-      origin: 'https://example.com',
-      pathname: '/blah/path/',
-      port: '',
-      protocol: 'https:',
-      search: '',
-    })
-    expect(isURLForDifferentApp('https://foo.com/blah/path/')).toBe(true)
-  })
-
-  it('[absolute URL] considers the apps the same if the URLs are both on the "newtab" subpath', () => {
-    const { isURLForDifferentApp } = require('../navigation')
-    setWindowLocation({
-      host: 'example.com',
-      hostname: 'example.com',
-      href: 'https://example.com/newtab/thing/',
-      origin: 'https://example.com',
-      pathname: '/newtab/thing/',
-      port: '',
-      protocol: 'https:',
-      search: '',
-    })
-    expect(isURLForDifferentApp('https://example.com/newtab/foobar/')).toBe(
-      false
-    )
-  })
-
-  it('[absolute URL] considers the apps the same if the URLs are both on the "search" subpath', () => {
-    const { isURLForDifferentApp } = require('../navigation')
-    setWindowLocation({
-      host: 'example.com',
-      hostname: 'example.com',
-      href: 'https://example.com/search/thing/',
-      origin: 'https://example.com',
-      pathname: '/search/thing/',
-      port: '',
-      protocol: 'https:',
-      search: '',
-    })
-    expect(isURLForDifferentApp('https://example.com/search/foobar/')).toBe(
-      false
-    )
-  })
-
-  it('[absolute URL] considers the apps different if the current location is on the "search" subpath and the URL is for the "homepage" app', () => {
-    const { isURLForDifferentApp } = require('../navigation')
-    setWindowLocation({
-      host: 'example.com',
-      hostname: 'example.com',
-      href: 'https://example.com/search/thing/',
-      origin: 'https://example.com',
-      pathname: '/search/thing/',
-      port: '',
-      protocol: 'https:',
-      search: '',
-    })
-    expect(isURLForDifferentApp('https://example.com/my-homepage/')).toBe(true)
-  })
-
-  it('[absolute URL] considers the apps different if the current location is on the "search" subpath and the URL is for the "newtab" app', () => {
-    const { isURLForDifferentApp } = require('../navigation')
-    setWindowLocation({
-      host: 'example.com',
-      hostname: 'example.com',
-      href: 'https://example.com/search/thing/',
-      origin: 'https://example.com',
-      pathname: '/search/thing/',
-      port: '',
-      protocol: 'https:',
-      search: '',
-    })
-    expect(isURLForDifferentApp('https://example.com/newtab/foobar/')).toBe(
-      true
-    )
-  })
-
-  it('[absolute URL] considers the apps different if the current location is on the "newtab" subpath and the URL is for the "homepage" app', () => {
-    const { isURLForDifferentApp } = require('../navigation')
-    setWindowLocation({
-      host: 'example.com',
-      hostname: 'example.com',
-      href: 'https://example.com/newtab/thing/',
-      origin: 'https://example.com',
-      pathname: '/newtab/thing/',
-      port: '',
-      protocol: 'https:',
-      newtab: '',
-    })
-    expect(isURLForDifferentApp('https://example.com/my-homepage/')).toBe(true)
-  })
-
-  it('[absolute URL] considers the apps different if the current location is on the "newtab" subpath and the URL is for the "search" app', () => {
-    const { isURLForDifferentApp } = require('../navigation')
-    setWindowLocation({
-      host: 'example.com',
-      hostname: 'example.com',
-      href: 'https://example.com/newtab/thing/',
-      origin: 'https://example.com',
-      pathname: '/newtab/thing/',
-      port: '',
-      protocol: 'https:',
-      search: '',
-    })
-    expect(isURLForDifferentApp('https://example.com/search/foobar/')).toBe(
-      true
-    )
-  })
-
-  it('[relative URL] considers the apps the same if the URL is exactly the same', () => {
-    const { isURLForDifferentApp } = require('../navigation')
-    setWindowLocation({
-      host: 'example.com',
-      hostname: 'example.com',
-      href: 'https://example.com/blah/path/',
-      origin: 'https://example.com',
-      pathname: '/blah/path/',
-      port: '',
-      protocol: 'https:',
-      search: '',
-    })
-    expect(isURLForDifferentApp('/blah/path/')).toBe(false)
-  })
-
-  it('[relative URL] considers the apps the same if the URLs are pages on the same "homepage" app (not part of the "newtab" or "search" subpaths)', () => {
-    const { isURLForDifferentApp } = require('../navigation')
-    setWindowLocation({
-      host: 'example.com',
-      hostname: 'example.com',
-      href: 'https://example.com/blah/path/',
-      origin: 'https://example.com',
-      pathname: '/blah/path/',
-      port: '',
-      protocol: 'https:',
-      search: '',
-    })
-    expect(isURLForDifferentApp('/some/thing/')).toBe(false)
-  })
-
-  it('[relative URL] considers the apps the same if the URLs are both on the "newtab" subpath', () => {
-    const { isURLForDifferentApp } = require('../navigation')
-    setWindowLocation({
-      host: 'example.com',
-      hostname: 'example.com',
-      href: 'https://example.com/newtab/thing/',
-      origin: 'https://example.com',
-      pathname: '/newtab/thing/',
-      port: '',
-      protocol: 'https:',
-      search: '',
-    })
-    expect(isURLForDifferentApp('/newtab/foobar/')).toBe(false)
-  })
-
-  it('[relative URL] considers the apps the same if the URLs are both on the "search" subpath', () => {
-    const { isURLForDifferentApp } = require('../navigation')
-    setWindowLocation({
-      host: 'example.com',
-      hostname: 'example.com',
-      href: 'https://example.com/search/thing/',
-      origin: 'https://example.com',
-      pathname: '/search/thing/',
-      port: '',
-      protocol: 'https:',
-      search: '',
-    })
-    expect(isURLForDifferentApp('/search/foobar/')).toBe(false)
-  })
-
-  it('[relative URL] considers the apps different if the current location is on the "search" subpath and the URL is for the "homepage" app', () => {
-    const { isURLForDifferentApp } = require('../navigation')
-    setWindowLocation({
-      host: 'example.com',
-      hostname: 'example.com',
-      href: 'https://example.com/search/thing/',
-      origin: 'https://example.com',
-      pathname: '/search/thing/',
-      port: '',
-      protocol: 'https:',
-      search: '',
-    })
-    expect(isURLForDifferentApp('/my-homepage/')).toBe(true)
-  })
-
-  it('[relative URL] considers the apps different if the current location is on the "search" subpath and the URL is for the "newtab" app', () => {
-    const { isURLForDifferentApp } = require('../navigation')
-    setWindowLocation({
-      host: 'example.com',
-      hostname: 'example.com',
-      href: 'https://example.com/search/thing/',
-      origin: 'https://example.com',
-      pathname: '/search/thing/',
-      port: '',
-      protocol: 'https:',
-      search: '',
-    })
-    expect(isURLForDifferentApp('/newtab/foobar/')).toBe(true)
-  })
-
-  it('[relative URL] considers the apps different if the current location is on the "newtab" subpath and the URL is for the "homepage" app', () => {
-    const { isURLForDifferentApp } = require('../navigation')
-    setWindowLocation({
-      host: 'example.com',
-      hostname: 'example.com',
-      href: 'https://example.com/newtab/thing/',
-      origin: 'https://example.com',
-      pathname: '/newtab/thing/',
-      port: '',
-      protocol: 'https:',
-      newtab: '',
-    })
-    expect(isURLForDifferentApp('/my-homepage/')).toBe(true)
-  })
-
-  it('[relative URL] considers the apps different if the current location is on the "newtab" subpath and the URL is for the "search" app', () => {
-    const { isURLForDifferentApp } = require('../navigation')
-    setWindowLocation({
-      host: 'example.com',
-      hostname: 'example.com',
-      href: 'https://example.com/newtab/thing/',
-      origin: 'https://example.com',
-      pathname: '/newtab/thing/',
-      port: '',
-      protocol: 'https:',
-      search: '',
-    })
-    expect(isURLForDifferentApp('/search/foobar/')).toBe(true)
   })
 })
